fix(ci): split databricks_borrowers proof helpers — green the file-size gate
`perf(sales): batch the /sales/aging borrower existence check` (db8dc9fd,
PR #170) added 48 lines to an 857-line file, pushing
backend/services/repositories/databricks_borrowers.py to 905 and over the
900-line FAIL threshold. Every CI run since #170 has failed on the
file-size gate; no other job was red.

Repo policy is to split, not to allowlist, so the self-contained proof unit
— the score-component vocabulary plus _build_borrower_proof and its
formula/offer-branch/evidence helpers — moves to a sibling
databricks_borrower_proof.py, matching the existing
databricks_genie_canonical/_direct precedent. The repository class only ever
touched that block through _build_borrower_proof, which is re-exported so
tests/fixtures importing it from the original path resolve to the same
function object.

905 -> 453 + 483. No behavior change.
